+++ NEWS.md
@@ -1,9 +1,11 @@
 # cfid 0.1.4
 
   * Fixed an issue related to fixed variables in the counterfactual graph.
-  * Fixed an issue related to counterfactual graph generation and equality of random variables.
+  * Fixed an issue related to counterfactual graph generation and equivalence of random variables.
   * Now uses lower case and snake case for classes.
   * Added `id` and `idc` algorithms for full identification pipeline. These algorithms can also be used directly via `causal_effect`.
+  * The syntax used by `dag` is now more flexible, allowing edges within subgraph definitions and nested subgraphs.
+  * Dropped dependency on R version 4.1.
   * Improved the package documentation.
 
 # cfid 0.1.3